+++ dipy/denoise/tests/meson.build
@@ -9,7 +9,6 @@ python_sources = [
   'test_noise_estimate.py',
   'test_non_local_means.py',
   'test_patch2self.py',
-  'test_qc.py',
   ]
