+++ dipy/stats/qc.py
