+++ dipy/stats/tests/meson.build
@@ -2,6 +2,7 @@ python_sources = [
   '__init__.py',
   'test_analysis.py',
   'test_resampling.py',
+  'test_qc.py',
   ]
