+++ dipy/stats/tests/test_qc.py
@@ -1,5 +1,5 @@
 import numpy as np
-from dipy.denoise.qc import neighboring_dwi_correlation
+from dipy.stats.qc import neighboring_dwi_correlation
 from dipy.core.gradients import gradient_table
 from dipy.core.geometry import normalized_vector
